+++ App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import DetailsScreen from './src/screens/Details';
-import GenderSelect from './src/screens/GenderSelect';
+import { GenderSelect } from './src/screens/GenderSelect/GenderSelect';
 
 type RootStackParamList = {
   Home: undefined;
